Further decoupling the contract verification
* In deploy-suite.js,
  - removed unused imports
  - added test env verification back in because it only checks addresses in the file against code being at those addresses in test env. Not an actual block explorer verification

* In hardhat.config.js,
  - added verify-suite task

* In package.json,
  - updated verify-suite scripts

* In polygon.verify.contracts.txt
  - updated the couple that actually verified previously

* Refactor / moved verify-suite.js to scripts/util
  - rewrote to be an async function that can be used in a hardhat task and passed arguments
