+++ hardhat.config.js
@@ -33,6 +33,18 @@ task("create-dispute-resolver", "Creates and activates a dispute resolver")
     await createAndActivateDR(path, createOnly, activateOnly);
   });
 
+task("verify-suite", "Verify contracts on the block explorer")
+  .addParam("chainId", "The chain id of the deployed contract address file")
+  .addParam("env", "The environment of the contract address file")
+  .setAction(async ({ chainId, env }) => {
+    const { verifySuite } = await lazyImport("./scripts/util/verify-suite");
+
+    // Contract list filter - empty array or use values from the name field of the contract object
+    // const filter = [];
+    const filter = ['GroupHandlerFacet','BosonVoucher Beacon', 'BosonVoucher Proxy'];
+    await verifySuite(chainId, env, filter);
+  });
+
 module.exports = {
   defaultNetwork: "hardhat",
   networks: {
